changed page not found logic, get id promise, added confirm dialog + some refactoring, added dialog rensponsive

+++ src/app/components/products/product-description/product-description.component.ts
@@ -76,6 +76,8 @@ export class ProductDescriptionComponent implements DoCheck {
   }
 
   deleteProduct() {
-    this.productService.deleteProduct(this.product);
+    if(confirm('Are you sure you want to delete ' + this.product.name + '?')) {
+      this.productService.deleteProduct(this.product);
+    }
   }
 }
